sbt-deploy: Make tasks reusable components of a deploy sequence
Each plugin exposes zero or more tasks that can be part of a deploy
sequence. The implementation of the actual deploy sequence if the
responsibility of the user. A plugin may also provide a set of methods,
for example, #scp and #ssh of SecureConnectivityPlugin.

In order to make sbt-deploy more user friendly, a basic deploy plugin
will be introduced at a later stage that merely deploys an init script
as well as the distribution while starting and stopping the service.

Also, the documentation is updated to reflect the API with examples.

Signed-off-by: Karim Osman <karim.osman@reaktor.fi>